Setup Trending Post Generation
When trending posts are called to be generated, the algorithm checks to see if there are left over trending posts in an array. If there are no more, 20 more are generated and are pushed to an array of posts that are trending. Each time a trending post is randomly chosen to be shown, the ones that are trending the most are pushed to a posts array. The post array is for all posts that are to be showed on the client side HTML page.